requirements.txt
anthropic
schedule
python-dotenv
flask
